IOS-3165 Move the suggestions to a separate file
